Dedicated file for CLI usage
