+++ package.json
@@ -5,7 +5,7 @@
   "main": "dist/src/index.js",
   "types": "dist/src/index.d.ts",
   "bin": {
-    "avro-to-typescript": "./dist/src/index.js"
+    "avro-to-typescript": "./dist/src/cli.js"
   },
   "directories": {
     "doc": "docs",
