+++ src/interfaces/CompilerOutputInterface.ts
@@ -1,0 +1,4 @@
+export interface CompilerOutputInterface {
+    class: string;
+    dir: string;
+}
